+++ src/search.ts
@@ -3,6 +3,7 @@ import parseQuery from "./parse-query";
 import filterCards from "./search-filters/cards";
 import filterColorIdentity from "./search-filters/color-identity";
 import filterComboData from "./search-filters/combo-data";
+import filterSize from "./search-filters/size";
 
 import type { SearchResults } from "./types";
 
@@ -32,6 +33,7 @@ export default async function search(query = ""): Promise<SearchResults> {
   combos = filterCards(combos, searchParams);
   combos = filterColorIdentity(combos, searchParams);
   combos = filterComboData(combos, searchParams);
+  combos = filterSize(combos, searchParams);
 
   return {
     errors,