+++ test/unit/search-filters/color-identity.test.ts
@@ -1,5 +1,4 @@
 import filterColorIdentity from "../../../src/search-filters/color-identity";
-import ColorIdentity from "../../../src/models/color-identity";
 import makeFakeCombo from "../../../src/make-fake-combo";
 
 import { mocked } from "ts-jest/utils";
@@ -44,8 +43,8 @@ describe("colorIdentityFilter", () => {
 
   describe("color filter", () => {
     it("can filter by color identity array with : operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is");
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is");
 
       params.colorIdentity.includeFilters.push({
         method: ":",
@@ -54,20 +53,20 @@ describe("colorIdentityFilter", () => {
 
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).not.toBeCalled();
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).not.toBeCalled();
 
       expect(result.length).toBe(1);
 
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
     });
 
     it("can filter by color identity array with >= operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "includes").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "includes").mockReturnValue(true);
 
       params.colorIdentity.includeFilters.push({
         method: ">=",
@@ -75,20 +74,20 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.includes).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.includes).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.includes).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.includes).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(1);
 
-      mocked(ColorIdentity.prototype.includes).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.includes).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
     });
 
     it("can filter by color identity array with > operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "includes").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "includes").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(false);
 
       params.colorIdentity.includeFilters.push({
         method: ">",
@@ -96,23 +95,23 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.includes).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.includes).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.includes).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.includes).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(1);
 
-      mocked(ColorIdentity.prototype.includes).mockReturnValue(false);
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.includes).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
     });
 
     it("can filter by color identity array with < operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(false);
 
       params.colorIdentity.includeFilters.push({
         method: "<",
@@ -120,22 +119,22 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(1);
 
-      mocked(ColorIdentity.prototype.isWithin).mockReturnValue(false);
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.isWithin).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
     });
 
     it("can filter by color identity array with <= operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
 
       params.colorIdentity.includeFilters.push({
         method: "<=",
@@ -143,19 +142,19 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(1);
 
-      mocked(ColorIdentity.prototype.isWithin).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.isWithin).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
     });
 
     it("can filter by color identity array with = operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(true);
 
       params.colorIdentity.includeFilters.push({
         method: "=",
@@ -163,12 +162,12 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(1);
 
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(0);
@@ -177,8 +176,8 @@ describe("colorIdentityFilter", () => {
 
   describe("negative color filter", () => {
     it("can filter out color identity array with : operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is");
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is");
 
       params.colorIdentity.excludeFilters.push({
         method: ":",
@@ -187,20 +186,20 @@ describe("colorIdentityFilter", () => {
 
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).not.toBeCalled();
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).not.toBeCalled();
 
       expect(result.length).toBe(0);
 
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
     });
 
     it("can filter out color identity array with >= operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "includes").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "includes").mockReturnValue(true);
 
       params.colorIdentity.excludeFilters.push({
         method: ">=",
@@ -208,20 +207,20 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.includes).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.includes).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.includes).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.includes).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(0);
 
-      mocked(ColorIdentity.prototype.includes).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.includes).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
     });
 
     it("can filter out color identity array with > operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "includes").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "includes").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(false);
 
       params.colorIdentity.excludeFilters.push({
         method: ">",
@@ -229,23 +228,23 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.includes).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.includes).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.includes).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.includes).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(0);
 
-      mocked(ColorIdentity.prototype.includes).mockReturnValue(false);
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.includes).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
     });
 
     it("can filter out color identity array with < operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(false);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(false);
 
       params.colorIdentity.excludeFilters.push({
         method: "<",
@@ -253,22 +252,22 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(0);
 
-      mocked(ColorIdentity.prototype.isWithin).mockReturnValue(false);
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.isWithin).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
     });
 
     it("can filter out color identity array with <= operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "isWithin").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "isWithin").mockReturnValue(true);
 
       params.colorIdentity.excludeFilters.push({
         method: "<=",
@@ -276,19 +275,19 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.isWithin).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.isWithin).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.isWithin).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(0);
 
-      mocked(ColorIdentity.prototype.isWithin).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.isWithin).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
     });
 
     it("can filter out color identity array with = operator", async () => {
-      jest.spyOn(ColorIdentity.prototype, "is").mockReturnValue(true);
+      jest.spyOn(combos[0].colorIdentity, "is").mockReturnValue(true);
 
       params.colorIdentity.excludeFilters.push({
         method: "=",
@@ -296,141 +295,15 @@ describe("colorIdentityFilter", () => {
       });
       let result = filterColorIdentity(combos, params);
 
-      expect(ColorIdentity.prototype.is).toBeCalledTimes(1);
-      expect(ColorIdentity.prototype.is).toBeCalledWith(["g", "r", "w"]);
+      expect(combos[0].colorIdentity.is).toBeCalledTimes(1);
+      expect(combos[0].colorIdentity.is).toBeCalledWith(["g", "r", "w"]);
 
       expect(result.length).toBe(0);
 
-      mocked(ColorIdentity.prototype.is).mockReturnValue(false);
+      mocked(combos[0].colorIdentity.is).mockReturnValue(false);
 
       result = filterColorIdentity(combos, params);
       expect(result.length).toBe(1);
-    });
-  });
-
-  describe("size filter", () => {
-    beforeEach(() => {
-      jest.spyOn(ColorIdentity.prototype, "numberOfColors");
-    });
-
-    it.each([":", "="])(
-      "can filter by color identity number of colors using %s",
-      async (operator) => {
-        mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(3);
-
-        params.colorIdentity.sizeFilters.push({
-          method: operator,
-          value: 3,
-        });
-
-        let result = await filterColorIdentity(combos, params);
-        expect(result.length).toBe(1);
-        expect(result[0].commanderSpellbookId).toBe("1");
-
-        mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(2);
-
-        result = filterColorIdentity(combos, params);
-        expect(result.length).toBe(0);
-
-        mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(4);
-
-        result = filterColorIdentity(combos, params);
-        expect(result.length).toBe(0);
-      }
-    );
-
-    it("can filter by color identity number of colors using >", async () => {
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(4);
-
-      params.colorIdentity.sizeFilters.push({
-        method: ">",
-        value: 3,
-      });
-
-      let result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(3);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(2);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
-    });
-
-    it("can filter by color identity number of colors using >=", async () => {
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(4);
-
-      params.colorIdentity.sizeFilters.push({
-        method: ">=",
-        value: 3,
-      });
-
-      let result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(3);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(2);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
-    });
-
-    it("can filter by color identity number of colors using <", async () => {
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(2);
-
-      params.colorIdentity.sizeFilters.push({
-        method: "<",
-        value: 3,
-      });
-
-      let result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(3);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(4);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
-    });
-
-    it("can filter by color identity number of colors using <=", async () => {
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(2);
-
-      params.colorIdentity.sizeFilters.push({
-        method: "<=",
-        value: 3,
-      });
-
-      let result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(3);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(1);
-      expect(result[0].commanderSpellbookId).toBe("1");
-
-      mocked(ColorIdentity.prototype.numberOfColors).mockReturnValue(4);
-
-      result = filterColorIdentity(combos, params);
-      expect(result.length).toBe(0);
     });
   });
 });